requirements.txt
flask==2.3.3
flask-wtf==1.1.1
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.5
python-dotenv==1.0.0
spotipy==2.22.1
google-api-python-client==2.88.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
authlib==1.2.1
google-auth-httplib2==0.1.0
pycountry==22.3.5
tqdm==4.65.0
pytest==7.3.1
gunicorn==20.1.0
flask-bcrypt==1.0.1
flask-login==0.6.2
flask-sqlalchemy==3.0.5
flask-session==0.5.0
email_validator==2.2.0
stripe==7.8.0
psycopg2-binary==2.9.9
# Flask-WTF 1.1 needs Werkzeug 2.x
werkzeug<3